# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OctoManger API v2
  version: 2.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v2/accounts/{id}:
    get:
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    patch:
      operationId: patchAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountPatchInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted account
...
  /api/v2/email/accounts/{id}/messages/latest:
    get:
      operationId: getLatestEmailMessage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: mailbox
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Latest message in the requested mailbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLatestMessageResult"
...
components:
  schemas:
...
    Account:
      type: object
      required:
        - id
        - identifier
        - status
        - tags
        - spec
      properties:
        id:
          type: integer
        account_type_id:
          type: integer
        account_type_key:
          type: string
        identifier:
          type: string
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        spec:
          type: object
          additionalProperties: true
...
    AccountPatchInput:
      type: object
      properties:
        status:
          type: string
        tags:
          type: array
          items:
            type: string
        spec:
          type: object
          additionalProperties: true
...
    EmailMessageDetail:
      type: object
      required:
        - id
        - subject
        - from
        - to
        - cc
        - date
        - size
        - flags
        - headers
        - text_body
        - html_body
      properties:
        id:
          type: string
        subject:
          type: string
        from:
          type: string
        to:
          type: string
        cc:
          type: string
        date:
          type: string
          format: date-time
        size:
          type: integer
        flags:
          type: array
          items:
            type: string
        headers:
          type: object
          additionalProperties:
            type: string
        text_body:
          type: string
        html_body:
          type: string
    EmailLatestMessageResult:
      type: object
      required:
        - mailbox
        - found
      properties:
        mailbox:
          type: string
        found:
          type: boolean
        item:
          $ref: "#/components/schemas/EmailMessageDetail"
...